database/requirements.txt
mysql-connector-python
Werkzeug
python-dotenv
